+++ twilio/internal/services/serverless/tests/resource_serverless_deployment_test.go
@@ -122,15 +122,6 @@ resource "twilio_serverless_build" "build" {
   function_version {
     sid = twilio_serverless_function.function.latest_version_sid
   }
-  dependencies = {
-    "twilio"                  = "3.6.3"
-    "fs"                      = "0.0.1-security"
-    "lodash"                  = "4.17.11"
-    "util"                    = "0.11.0"
-    "xmldom"                  = "0.1.27"
-    "@twilio/runtime-handler" = "1.0.1"
-  }
-
   polling {
     enabled = true
   }