+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.8.0 (unreleased)
+
+FEATURES
+
+- **Updated Resource:** `twilio_serverless_build` Add `runtime` argument and `runtime` attribute
+- **Updated Data Source:** `twilio_serverless_build` Add `runtime` attribute
+- **Updated Data Source:** `twilio_serverless_builds` Add `runtime` attribute
+
 ## v0.7.1 (2021-02-08)
 
 Update conversation documentation to remove typos and improve the readability of the documentation